+++ azd/src/main/java/org/azd/enums/Instance.java
@@ -24,4 +24,8 @@ public enum Instance {
     public String getInstance() {
         return instance;
     }
+
+    public String setSubdomain(String subdomainName) {
+        return instance.replaceFirst("https://", "https://" + subdomainName + ".");
+    }
 }