+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 ## 5.0.3
 
 - Added support for GitRefs, GitTag in **GitApi**.